SLW: Enable Stop4 as the deepest idle state
Adding stop4 to device tree as the deepest idle state for power9 dd1.
The latency is configured such that they are not chosen for cpuidle and
will only be used for cpuhotplug.

Use a common variable has_wakeup_engine instead of has_slw to tell if
the
a) SLW image is populated in case of power8
b) CME image is populated in case of power9

Currently we expect CME to be loaded if homer address is known ( except
for simulators)

Signed-off-by: Akshay Adiga <akshay.adiga@linux.vnet.ibm.com>